+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-mobile-spec.js
@@ -16,17 +16,18 @@ describe('Date Time Picker Twelve Hour Mobile', () => {
   describe('Date Time Picker Twelve Hour Mobile - Default with no time', () => {
     before(() => browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-12-hour-mobile'));
 
-    Terra.should.validateElement('default', { axeRules: { rules: ignoredA11y } });
+    Terra.it.validatesElement('default', { axeRules: { rules: ignoredA11y } });
   });
 
   describe('Date Time Picker Twelve Hour Mobile - Clicking Meridiem updates the time', () => {
-    before(() => {
+    it('Date Time Picker Twelve Hour Mobile - Clicking Meridiem updates the time', () => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-12-hour-mobile');
 
       browser.click('//*[@id="datetimeValueProvided"]/div[2]/div/div[3]/button[1]');
-    });
+      Terra.validates.element('default', { axeRules: { rules: ignoredA11y } });
 
-    Terra.should.validateElement('default', { axeRules: { rules: ignoredA11y } });
+      browser.click('//*[@id="datetimeValueProvided"]/div[2]/div/div[3]/button[2]');
+    });
   });
 
   describe('Date Time Picker Twelve Hour Mobile - Focus Hour Styles', () => {
@@ -36,7 +37,7 @@ describe('Date Time Picker Twelve Hour Mobile', () => {
       browser.click('//*[@id="terra-time-hour-input"]');
     });
 
-    Terra.should.validateElement('default', { axeRules: { rules: ignoredA11y } });
+    Terra.it.validatesElement('default', { axeRules: { rules: ignoredA11y } });
   });
 
   describe('Date Time Picker Twelve Hour Mobile - Focus Minute Styles', () => {
@@ -46,6 +47,6 @@ describe('Date Time Picker Twelve Hour Mobile', () => {
       browser.click('//*[@id="terra-time-minute-input"]');
     });
 
-    Terra.should.validateElement('default', { axeRules: { rules: ignoredA11y } });
+    Terra.it.validatesElement('default', { axeRules: { rules: ignoredA11y } });
   });
 });