+++ packages/terra-date-picker/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
 * Replaced terra-hookshot with terra-popup
 
 4.2.0 - (May 21, 2019)